Make storage copying universal
