--- src/commonTest/kotlin/qbit/MemStorageTest.kt
@@ -1,14 +1,0 @@
-package qbit
-
-import qbit.storage.MemStorage
-import kotlin.test.Test
-
-
-class MemStorageTest : StorageTest() {
-
-    @Test
-    fun testMemStorage() {
-        testStorage(MemStorage())
-    }
-
-}
