+++ src/commonTest/kotlin/qbit/storage/MemStorageTest.kt
@@ -1,0 +1,12 @@
+package qbit
+
+import qbit.storage.MemStorage
+import qbit.storage.Storage
+
+class MemStorageTest : StorageTest() {
+
+    override fun storage(): Storage {
+        return MemStorage()
+    }
+
+}
