+++ src/commonTest/kotlin/qbit/storage/StorageTest.kt
@@ -3,16 +3,19 @@ package qbit
 import qbit.ns.Namespace
 import qbit.ns.ns
 import qbit.ns.root
-import qbit.platform.Files
-import qbit.storage.FileSystemStorage
 import qbit.storage.MemStorage
 import qbit.storage.Storage
+import qbit.storage.copyStorage
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-open class StorageTest {
+abstract class StorageTest {
 
-    protected fun testStorage(storage: Storage) {
+    abstract fun storage(): Storage
+
+    @Test
+    fun testStorage() {
+        val storage = storage()
         val rootBytes = byteArrayOf(0, 0, 0, 0)
         val subBytes = byteArrayOf(1, 1, 1, 1)
         val rootNs = Namespace("test-root")
@@ -26,6 +29,20 @@ open class StorageTest {
 
         assertArrayEquals(subBytes, storage.load(subNs["sub-data"]))
         assertEquals(setOf(subNs["sub-data"]), storage.keys(subNs).toSet())
+    }
+
+    @Test
+    fun testCopyNsConstructor() {
+        val testNs = ns("nodes")("test")
+
+        val origin = MemStorage()
+        setupTestSchema(origin)
+
+        // actually it compiles
+        val storage = storage()
+        copyStorage(origin, storage)
+        assertEquals(origin.subNamespaces(testNs.parent!!), storage.subNamespaces(testNs.parent!!))
+        assertEquals(storage.subNamespaces(root).sortedBy { it.name }, listOf(ns("nodes"), ns("refs")).sortedBy { it.name })
     }
 
 }
